+++ package.json
@@ -54,6 +54,7 @@
     "abort-controller": "^3.0.0",
     "aegir": "^28.1.0",
     "bl": "^4.0.0",
+    "is-loopback-addr": "^1.0.1",
     "it-goodbye": "^2.0.1",
     "it-pipe": "^1.0.1",
     "libp2p-interfaces": "^0.7.1",
